Cleanup: remove redundant and unnecessary casts
Drop casts that the compiler already provides implicitly or that are no-ops:
- redundant numeric casts around Math.Round / division / TimeSpan.FromSeconds
  where the operand or result type already matches (integer division semantics
  in AdvancedEffectAudioTextPulse are preserved);
- pointless (uint)/(int)/(byte) casts on already-correct types;
- (IEnumerable<T>) upcasts left over before LINQ instance calls;
- superfluous nullable/reference casts on ternary branches (e.g. : (T?)null).

No behavior change; full solution builds clean.

+++ src/libse/Common/ColorUtils.cs
@@ -44,12 +44,12 @@ namespace Nikse.SubtitleEdit.Core.Common
 
         public static uint ToArgb(this SKColor color)
         {
-            var number = (uint)color.Alpha << 24 | (uint)color.Red << 16 | (uint)color.Green << 8 | (uint)color.Blue;
+            var number = (uint)color.Alpha << 24 | (uint)color.Red << 16 | (uint)color.Green << 8 | color.Blue;
             return number;
         }
         public static string ToArgbString(this SKColor color)
         {
-            var number = (uint)color.Alpha << 24 | (uint)color.Red << 16 | (uint)color.Green << 8 | (uint)color.Blue;
+            var number = (uint)color.Alpha << 24 | (uint)color.Red << 16 | (uint)color.Green << 8 | color.Blue;
             return number.ToString(CultureInfo.InvariantCulture);
         }
