+++ src/ui/Features/Files/ExportImageBased/ExportImageBasedViewModel.cs
@@ -730,7 +730,7 @@ public partial class ExportImageBasedViewModel : ObservableObject
             ip.Alignment == ExportAlignment.MiddleLeft ||
             ip.Alignment == ExportAlignment.BottomLeft)
         {
-            x = (int)ip.LeftRightMargin;
+            x = ip.LeftRightMargin;
         }
         else if (ip.Alignment == ExportAlignment.TopCenter ||
                  ip.Alignment == ExportAlignment.MiddleCenter ||
@@ -749,7 +749,7 @@ public partial class ExportImageBasedViewModel : ObservableObject
             ip.Alignment == ExportAlignment.TopCenter ||
             ip.Alignment == ExportAlignment.TopRight)
         {
-            y = (int)ip.BottomTopMargin;
+            y = ip.BottomTopMargin;
         }
         else if (ip.Alignment == ExportAlignment.MiddleLeft ||
                  ip.Alignment == ExportAlignment.MiddleCenter ||
