+++ src/ui/Features/Video/BlankVideo/BlankVideoViewModel.cs
@@ -285,7 +285,7 @@ public partial class BlankVideoViewModel : ObservableObject
         if (long.TryParse(arr[1].Trim(), out var f))
         {
             _processedFrames = f;
-            ProgressValue = (double)_processedFrames * 100.0 / JobItems[_jobItemIndex].TotalFrames;
+            ProgressValue = _processedFrames * 100.0 / JobItems[_jobItemIndex].TotalFrames;
         }
     }
